inference/requirements.txt
numpy==1.19.5
opencv-python==4.5.5.64
transformers==4.16.2
diffusers==0.2.4
tqdm==4.64.0
openvino==2022.3.0
huggingface_hub==0.9.0
scipy==1.9.0
streamlit==1.12.0
watchdog==2.1.9
ftfy==6.1.1
streamlit_drawable_canvas==0.9.1
pillow==9.0.1
piexif==1.1.3
python-magic
cnvrg